+++ odtbrain/_Back_3D_tilted.py
@@ -5,6 +5,7 @@ from __future__ import division, print_function
 import ctypes
 import gc
 import multiprocessing as mp
+import numexpr as ne
 import numpy as np
 from numpy import cos, sin
 
@@ -378,7 +379,8 @@ def backpropagate_3d_tilted(uSin, angles, res, nm, lD=0,
                      offset_alpha=0, offset_beta=0,
                      padding=(True, True), padfac=1.75, padval=None,
                      intp_order=2, dtype=_np_float64,
-                     num_cores=_ncores, 
+                     num_cores=_ncores,
+                     use_numexpr=True, 
                      jmc=None, jmm=None,
                      verbose=_verbose):
     u""" 3D backpropagation with the Fourier diffraction theorem
@@ -491,6 +493,8 @@ def backpropagate_3d_tilted(uSin, angles, res, nm, lD=0,
     num_cores : int
         The number of cores to use for parallel operations. This value
         defaults to the number of cores on the system.
+    use_numexpr : bool
+        Use numexpr for fast evaulation of expressions.
     jmc, jmm : instance of :func:`multiprocessing.Value` or ``None``
         The progress of this function can be monitored with the 
         :mod:`jobmanager` package. The current step ``jmc.value`` is
@@ -841,13 +845,17 @@ def backpropagate_3d_tilted(uSin, angles, res, nm, lD=0,
     # simply executing:
     # filter2 = np.exp(1j * zv * km * (Mp - 1))
     
-    Mpm1 = km * (Mp - 1)
-    args=zip(zv.flatten(), [Mpm1]*zv.shape[0])
-    filter2_pool = mp.Pool(processes=num_cores)
-    filter2 = filter2_pool.map(_filter2_func, args)
-    filter2_pool.terminate()
-    filter2_pool.terminate()
-    del filter2_pool, args, Mpm1
+    if use_numexpr:
+        ne.set_num_threads(num_cores)
+        filter2 = ne.evaluate("exp(1j * zv * km * (Mp - 1))")
+    else:
+        Mpm1 = km * (Mp - 1)
+        args=zip(zv.flatten(), [Mpm1]*zv.shape[0])
+        filter2_pool = mp.Pool(processes=num_cores)
+        filter2 = filter2_pool.map(_filter2_func, args)
+        filter2_pool.terminate()
+        filter2_pool.terminate()
+        del filter2_pool, args, Mpm1
 
     # occupies some amount of ram
     #filter2[0].size*len(filter2)*128/8/1024**3
@@ -858,7 +866,6 @@ def backpropagate_3d_tilted(uSin, angles, res, nm, lD=0,
     #                               a, z, y,  x
     #projection = projection.reshape(la, 1, lNy, lNx)
     projection = projection.reshape(la, lNy, lNx)
-
 
     # This frees comparatively few data
     del M
